refactor(frontend): separate component for container inspect
+ added separate component for container inspect tab
+ minor style fixes
