+++ frontend/src/app/features/container-card/container-card-logs/container-card-logs.scss
@@ -18,6 +18,6 @@
   }
 
   .logs-text {
-    overflow-y: auto;
+    overflow-x: auto;
   }
 }
